Develop (#29)
* Update index.md

principlal_idns="my_principal_idns" 
should this be principal_idns="my_principal_idns" 
(left as is)

* Update index.md

The `MetadataSession` is authenticated using the `WorldcatAccessToken` object. The session allows searching brief records as well as retrieving full bibs in the MARX XML format.
or 
The `MetadataSession` is authenticated using the `WorldcatAccessToken` object. The session allows searching brief records as well as retrieving full bibs in the MARC XML format.
(changed to MARC)